requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
python-dotenv==1.2.2
psycopg==3.3.3
PyJWT==2.12.0
sqlparse==0.5.5
gunicorn==23.0.0
openpyxl==3.1.5
whitenoise==6.9.0
dj-database-url==2.3.0
